test: Todos os testes de cenários do CRUD de playlist realizados

+++ backend/src/services/playlist.service.ts
@@ -7,6 +7,9 @@ class PlaylistServiceMessageCode {
     public static readonly playlist_not_found = 'playlist_not_found';
     public static readonly playlist_update_error = 'playlist_update_error';
     public static readonly playlist_creation_error = 'playlist_creation_error';
+    public static readonly playlist_deletion_error = 'playlist_deletion_error';
+    public static readonly add_song_error = "add_song_error";
+    public static readonly delete_song_error = "delete_song_error";
 }
 
 
@@ -21,8 +24,6 @@ class PlaylistService {
     const playlistEntity = await this.playlistRepository.getPlaylists()
 
     const playlistsModel = playlistEntity.map((play) => new PlaylistModel(play));
-
-    //Adicionar um erro?
 
     return playlistsModel;
   }
@@ -43,9 +44,18 @@ class PlaylistService {
   }
 
   public async deletePlaylistById(id: string): Promise<void>{
+
+    const playlistEntity = await this.playlistRepository.getPlaylistById(id);
+
+    if(!playlistEntity){
+      throw new NotFoundError({
+        msg: 'Playlist id not found',
+        msgCode: PlaylistServiceMessageCode.playlist_deletion_error,
+      });
+    }
+
     await this.playlistRepository.deletePlaylistById(id);
 
-    //Adicionar um erro?
   }
   
   public async updatePlaylistById(newPlaylist: PlaylisEntity, id: string): Promise<PlaylistModel>{
@@ -64,6 +74,14 @@ class PlaylistService {
   }
 
   public async createPlaylist(newPlaylist: PlaylisEntity): Promise<PlaylistModel>{
+
+    if(!newPlaylist.name){
+      throw new InternalServerError({
+        msg: 'name field empty',
+        msgCode: PlaylistServiceMessageCode.playlist_creation_error,
+      });
+    }
+
     const playlistEntity = await this.playlistRepository.createPlaylist(newPlaylist);
     
     if(!playlistEntity){
@@ -79,16 +97,54 @@ class PlaylistService {
   }
 
   public async deleteSongUsingPlaylistId(song: string, id: string): Promise<void>{
-    await this.playlistRepository.deleteSongUsingPlaylistId(song, id);
+    const playlistEntity = await this.playlistRepository.getPlaylistById(id);
 
-    //add erro?
+    if(!playlistEntity){
+      throw new NotFoundError({
+        msg: 'Playlist id not found',
+        msgCode: PlaylistServiceMessageCode.playlist_not_found,
+      });
+    }
+    if(!playlistEntity.songs.includes(song)){
+      throw new NotFoundError({
+        msg: 'Song not in playlist',
+        msgCode: PlaylistServiceMessageCode.delete_song_error,
+      });
+
+    }
+    if(!song){
+      throw new InternalServerError({
+        msg: 'empty song name',
+        msgCode: PlaylistServiceMessageCode.delete_song_error,
+      });
+    }
+
+    await this.playlistRepository.deleteSongUsingPlaylistId(song, id);
   }
 
   public async addSongUsingPlaylistId(song: string, id: string): Promise<void>{
+    const playlistEntity = await this.playlistRepository.getPlaylistById(id);
+    if(!playlistEntity){
+      throw new NotFoundError({
+        msg: 'Playlist id not found',
+        msgCode: PlaylistServiceMessageCode.playlist_not_found,
+      });
+    }
+    if(playlistEntity.songs.includes(song)){
+      throw new InternalServerError({
+        msg: 'Song already in playlist',
+        msgCode: PlaylistServiceMessageCode.delete_song_error,
+      });
+
+    }
+    if(!song){
+      throw new InternalServerError({
+        msg: 'empty song name',
+        msgCode: PlaylistServiceMessageCode.add_song_error,
+      });
+    }
+    
     await this.playlistRepository.addSongUsingPlaylistId(song, id);
-
-    //add erro?
   }
-
 }
 export default PlaylistService;
